Refactors code to conform to Maven Standard Directory structure; Adds module-info.java; Closes #1
