+++ src/main/java/termbank/model/DataBank.java
