+++ src/main/java/termbank/view/BankSheetController.java
@@ -27,7 +27,7 @@ import javafx.scene.control.TableView;
 import javafx.scene.control.TextArea;
 import javafx.scene.control.TextField;
 import javafx.scene.control.cell.TextFieldTableCell;
-import termbank.AlertHelper;
+import termbank.utils.AlertHelper;
 import termbank.TermBankApp;
 import termbank.model.DataBank;
